Create KaratCodingQuestion.js
I had just one edge case that was causing me to fail my coding question. I solved it about five minutes after the interview ended. Here is a Github link to the working code (I didn't refactor since I wanted to prove, based on the time stamp when I submit this, that the code is working.

Please look at the timestamp when I committed this code file and notice that it is directly after my call ended.